~correcting my intuition about c-strings
On converting, I would sometimes get the other tag missing. This was
both on the repl and the tests, and was quite annoying.

Each c-string should end in a null terminator, which I forgot about. So
this commit fixes that by setting the last character pointer on each
branch to the null terminator.