~compile_lines -> compile_file, shuffle
I've redid the compiler to use a FILE object pointer, which means that
we can leave error checking for FILES on the main program.

Furthermore, I've removed the use of the array_t type with a two pass
system. One pass is done for the number of lines (along with potentially
other diagnostics), another is done for the actual compilation. This
means at runtime we have an understanding about the size of the file and
can optimize for it. Also, we have a single memory block to assign to,
where we know the size of the array.