-list item regex, +nitems variable, ~bold and italic
List item regex can be replaced with a simple algorithm to check the
first character of the line to see if it's a bullet character.

I must check the bold before the italic code, as the regex may greedily
eat up the double asterisks if I don't use it.

squash! -list item regex, +nitems variable